Improve type handling of field length

+++ src/types.rs
@@ -75,7 +75,7 @@ pub struct FieldDefinition {
     pub number: u8,
     /// The length of the field in bytes. This may be a whole multiple of the intrinsic length of
     /// the field's data-type, which indicates field is an array of the base data-type.
-    pub length: u8,
+    pub length: usize,
     /// A bit-mask identifying the data-type of the field.
     ///
     /// This implementation only looks at the lower 4 bits. The upper 4 bits contain redundant
